pyproject.toml
[tool.poetry]
package-mode = false  

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.12"
uvicorn = {extras = ["standart"], version = "^0.34.0"}
pydantic = {extras = ["email"], version = "^2.11.1"}
pydantic-settings = "^2.8.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.40"}
asyncpg = "^0.30.0"
aiosmtplib = "^4.0.0"
python-multipart = "^0.0.20"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
bcrypt = "^4.3.0"
faststream = {extras = ["cli"], version = "^0.5.37"}
alembic = "^1.15.2"
redis = "^5.2.1"
itsdangerous = "^2.2.0"
nats-py = "^2.10.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

